requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
passlib
PyJWT
pydantic[email]